+++ src/model/table/viewNode.ts
@@ -3,8 +3,13 @@ import * as vscode from "vscode";
 import { TableNode } from "./tableNode";
 import { DatabaseCache } from "../../database/DatabaseCache";
 import { ModelType, Constants } from "../../common/Constants";
+import { ConnectionManager } from "../../database/ConnectionManager";
+import { QueryUnit } from "../../database/QueryUnit";
+import { MySQLTreeDataProvider } from "../../provider/MysqlTreeDataProvider";
 
 export class ViewNode extends TableNode{
+    type:string=ModelType.VIEW;
+
     public getTreeItem(): vscode.TreeItem {
 
         this.identify = `${this.host}_${this.port}_${this.user}_${this.database}_${this.table}`
@@ -19,6 +24,22 @@ export class ViewNode extends TableNode{
                 arguments: [this, true]
             }
         };
+    }
+
+    public drop() {
+
+        vscode.window.showInputBox({ prompt: `Are you want to drop view ${this.table} ?     `, placeHolder: 'Input y to confirm.' }).then(async inputContent => {
+            if (inputContent.toLocaleLowerCase() == 'y') {
+                QueryUnit.queryPromise(await ConnectionManager.getConnection(this), `DROP view ${this.database}.${this.table}`).then(() => {
+                    DatabaseCache.clearTableCache(`${this.host}_${this.port}_${this.user}_${this.database}`)
+                    MySQLTreeDataProvider.refresh()
+                    vscode.window.showInformationMessage(`Drop view ${this.table} success!`)
+                })
+            } else {
+                vscode.window.showInformationMessage(`Cancel drop view ${this.table}!`)
+            }
+        })
 
     }
+
 }